16-bit move; boardLookup; no validation; bit opps

+++ src/main/java/view/board/PieceGraphic.java
@@ -11,7 +11,8 @@ import model.GameInfo;
 public class PieceGraphic {
     /*
      * Piece ennumeration order is based on the order in pieces.png to make it easier to set
-     * the image coordinates in setImageCoords()
+     * the image coordinates in setImageCoords(). Note this enumeration is NOT the same as
+     * the enumeration used in BoardLookup
      * 
      * None: 0
      * King: 1
@@ -61,7 +62,7 @@ public class PieceGraphic {
         );
     }
 
-    public static void addFadedPieceToStack(double mouseX, double mouseY) {
+    public static void addFadedPieceToStack(float mouseX, float mouseY) {
         setImageCoords(
             Bitboard.getPieceIDFromKey(
                 Bitboard.getKeyFromBitIndex(
@@ -95,7 +96,7 @@ public class PieceGraphic {
         pieceCanvasContext.clearRect(0, 0, GameInfo.getBoardLength(), GameInfo.getBoardLength());
     }
 
-    public static void clearByMouse(double mouseX, double mouseY) {
+    public static void clearByMouse(float mouseX, float mouseY) {
         Vec2 clearSquareUV = Convert.mouseToCorner(mouseX, mouseY);
 
         pieceCanvasContext.clearRect(
